fix(visualization): convert float to integer in datetime indexing
- Fix "cannot do positional indexing on DatetimeIndex with these indexers of type float" error
- Convert result of hours_to_show and days_to_show calculations to integer
- Ensure plot_data.tail() receives an integer value for the number of candles
- Maintain correct display of candlesticks and volume when using tail() method

Resolves issue with DatetimeIndex requiring integer indexers instead of float values.

+++ src/visualization/chart_generator.py
@@ -125,10 +125,11 @@ class ChartGenerator:
         if not isinstance(df_copy.index, pd.DatetimeIndex):
             df_copy.index = pd.to_datetime(df_copy.index)
         
-        # Trim data to the requested time range - DŮLEŽITÉ: POUŽIJ POSLEDNÍCH N SVÍČEK MÍSTO ČASOVÉHO ROZSAHU!
+        # Trim data to the requested time range - POUŽIJ POSLEDNÍCH N SVÍČEK MÍSTO ČASOVÉHO ROZSAHU
         if hours_to_show:
             # Pro intraday použijme fixní počet svíček místo času
-            num_candles = min(hours_to_show * 2, len(df_copy))  # Přibližný odhad - 2 svíčky na hodinu
+            # Oprava - používáme integer místo float
+            num_candles = min(int(hours_to_show * 2), len(df_copy))  # Přibližný odhad - 2 svíčky na hodinu
             plot_data = df_copy.tail(num_candles).copy()
         else:
             # Pro denní grafy použijeme fixní počet svíček odpovídající přibližně počtu dnů
@@ -142,7 +143,8 @@ class ChartGenerator:
             elif timeframe == '30m':
                 candles_per_day = 48
             
-            num_candles = min(days_to_show * candles_per_day, len(df_copy))
+            # Oprava - používáme integer místo float
+            num_candles = min(int(days_to_show * candles_per_day), len(df_copy))
             plot_data = df_copy.tail(num_candles).copy()
         
         # Nastavení parametrů pro mplfinance
